Added a new sprite sheet, new class, new objects to game
Sprite sheet is from an RPG tileset maker called Stealthix.  New class is called RiverLog to represent the log objects in the game.  New objects are the river logs (6 of them) which move from the left and right side of the screen.  Frog can jump on each log, but must work on what happens if the log reaches edge of the screen (boundaries).