Made collision code changes and started drawing a new menu screen
Finished implementing the correct collision code between the frog character and the goal posts (frog will now land on each goal post correctly and gain the correct points and additional points if there is a dragonfly.) If the frog touches the rocks or the waters in the top row then it will lose a life and be sent back to the start! I also started drawing a concept for a high scores screen which I will implement next time (hopefully its not super tricky!) then after that I only need to test the game out and it should be done.