Added two new functions to the game code for high scores
    8-11-21 Started work on the game screen that displays personal score of the player after the
    game ends and allows for the player to enter their custom name in a text box.  The score and name are saved to a file of
high scores and the player names. Also created the additional game screen which shows a high score screen after the
player runs out of lives and this screen is displayed before the game over screen and after the personal score screen.
Using the high scores text file, it displays  several scores along with several names including the players name and the
players score which were captured in the enter personal score screen.